Add .gitattributes and introduce end-of-line normalization
This patch sets text=auto in the .gitattributes file
and which ensures that line endings are normalized.
Moreover all of the files that are currently use DOS
line endings will remain this way.

As of this patch any new file that will be added to
this repository should use UNIX line endings.

Signed-off-by: Dmitry Fleytman <dmitry@daynix.com>
Signed-off-by: Sameeh Jubran <sameeh@daynix.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>